Replaced OBSOutputInfo
